Reimplement QAbstractSocket::setReadBufferSize() properly
This function takes a qint64 argument, not an int argument.

peerwireclient.h:119:10: warning: 'PeerWireClient::setReadBufferSize'
     hides overloaded virtual function [-Woverloaded-virtual]
    void setReadBufferSize(int size);
         ^
src/network/socket/qabstractsocket.h:162:18: note: hidden overloaded
     virtual function 'QAbstractSocket::setReadBufferSize' declared here
    virtual void setReadBufferSize(qint64 size);
                 ^

+++ examples/network/torrent/peerwireclient.cpp
@@ -375,7 +375,7 @@ qint64 PeerWireClient::uploadSpeed() const
     return sum / (8 * 2);
 }
 
-void PeerWireClient::setReadBufferSize(int size)
+void PeerWireClient::setReadBufferSize(qint64 size)
 {
     socket.setReadBufferSize(size);
 }
